requirements.txt
grpcio==1.50.0
grpcio-tools==1.50.0
protobuf==4.21.9
psycopg2==2.9.5
six==1.16.0
